# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoChat 聊天/好友/房间管理后端接口",
        "title": "GoChat Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/friendship-requests/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户相关的好友请求列表，可基于 base_id 游标和 limit 分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendship"
                ],
                "summary": "获取好友请求列表",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "分页游标，返回该ID之前的记录",
                        "name": "base_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页大小，默认20，最大100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回好友请求列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_friendship_port_http.ListFriendRequestsResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向指定用户发送好友请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendship"
                ],
                "summary": "发送好友请求",
                "parameters": [
                    {
                        "description": "发送好友请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_friendship_port_http.SendFriendRequestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或业务校验失败",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
...
        "/notifications/system-messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的系统通知列表，可基于 base_id 游标和 limit 分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "获取系统通知列表",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "分页游标，返回该ID之前的消息",
                        "name": "base_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页大小，默认20，最大100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回系统通知列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_notification_port_http.ListSystemMessagesResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
        "/profiles/me": {
            "get": {
                "description": "根据token获取本用户资料",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "获取本用户资料",
                "responses": {
                    "200": {
                        "description": "成功返回用户资料",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_profile_port_http.GetMyProfileResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前登录用户的基础资料（昵称、邮箱、电话等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "更新当前用户资料",
                "parameters": [
                    {
                        "description": "更新用户资料请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_profile_port_http.UpdateUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "404": {
                        "description": "用户资料不存在",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
...
        "/profiles/users/{user_id}": {
            "get": {
                "description": "根据用户ID获取用户公开资料",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "获取用户资料",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回用户资料",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_profile_port_http.GetUserProfileResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gochat_internal_friendship_domain.FriendRequestState": {
            "type": "string",
            "enum": [
                "pending",
                "agreed",
                "refused"
            ],
            "x-enum-varnames": [
                "StatePending",
                "StateAgreed",
                "StateRefused"
            ]
        },
...
        "gochat_internal_friendship_dto.FriendRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "from": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                },
                "id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.OperationID"
                },
                "sent_at": {
                    "description": "UTC",
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/gochat_internal_friendship_domain.FriendRequestState"
                },
                "to": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                }
            }
        },
...
        "gochat_internal_notification_domain.MessageState": {
            "type": "string",
            "enum": [
                "undelivered",
                "delivered"
            ],
            "x-enum-varnames": [
                "MessageStateUndelivered",
                "MessageStateDelivered"
            ]
        },
        "gochat_internal_notification_dto.SystemMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.MessageID"
                },
                "recipient_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                },
                "sent_at": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/gochat_internal_notification_domain.MessageState"
                }
            }
        },
...
        "gochat_internal_profile_dto.UserProfile": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.Gender"
                },
                "id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "sign": {
                    "type": "string"
                },
                "signed_up_at": {
                    "type": "string"
                }
            }
        },
...
        "gochat_internal_shared_api.Code": {
            "type": "integer",
            "enum": [
                200,
                500,
                400,
                401,
                402,
                403,
                404
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeServerError",
                "CodeTimeout",
                "CodeInvalidParam",
                "CodeInvalidToken",
                "CodeNotFound",
                "CodeServiceUnavailable"
            ]
        },
        "gochat_internal_shared_api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/gochat_internal_shared_api.Code"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "gochat_internal_shared_kernel.Gender": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "UnknownGender",
                "MaleGender",
                "FemaleGender"
            ]
        },
        "gochat_internal_shared_kernel.MessageID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
        "gochat_internal_shared_kernel.OperationID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
...
        "gochat_internal_shared_kernel.UserID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
...
        "internal_friendship_port_http.ListFriendRequestsResponseData": {
            "type": "object",
            "properties": {
                "requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gochat_internal_friendship_dto.FriendRequest"
                    }
                }
            }
        },
...
        "internal_friendship_port_http.SendFriendRequestRequest": {
            "type": "object",
            "required": [
                "content",
                "to_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 100
                },
                "to_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                }
            }
        },
        "internal_notification_port_http.ListSystemMessagesResponseData": {
            "type": "object",
            "properties": {
                "system_messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gochat_internal_notification_dto.SystemMessage"
                    }
                }
            }
        },
        "internal_profile_port_http.GetMyProfileResponseData": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/gochat_internal_profile_dto.UserProfile"
                }
            }
        },
...
        "internal_profile_port_http.GetUserProfileResponseData": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/gochat_internal_profile_dto.UserProfile"
                }
            }
        },
...
        "internal_profile_port_http.UpdateUserProfileRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.Gender"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "sign": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}